requirements.txt
streamlit
pytesseract
PyMuPDF
pillow
opencv-python
numpy
pandas
openai
PyGithub